backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.7.0
werkzeug>=3.0.0
openai>=1.0.0
pypdf2>=3.0.0
# Optional:
gunicorn>=22.0.0
# psycopg2-binary    (for PostgreSQL: pip install psycopg2-binary)
